backend/pyproject.toml
[project]
name = "utah-healthcare-workforce"
version = "0.1.0"
description = "Utah Healthcare Workforce Intelligence Dashboard — backend API and forecasting engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "fredapi>=0.5.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "statsmodels>=0.14.0",
    "prophet>=1.1.5",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
